+++ src/models/form.model.ts
@@ -7,6 +7,8 @@ import { layoutSchema } from './layout.model';
 import { Version } from './version.model';
 import { Record } from './record.model';
 import { getGraphQLTypeName } from '@utils/validators';
+import { deleteFolder } from '@utils/files/deleteFolder';
+import { logger } from '@services/logger.service';
 
 /** Form documents interface declaration */
 interface FormDocument extends Document {
@@ -161,10 +163,19 @@ schema.statics.hasDuplicate = function (
 
 // handle cascading deletion for forms
 addOnBeforeDeleteMany(schema, async (forms) => {
-  const versions = forms.reduce((acc, form) => acc.concat(form.versions), []);
-  await Record.deleteMany({ form: { $in: forms } });
-  await Channel.deleteMany({ form: { $in: forms } });
-  await Version.deleteMany({ _id: { $in: versions } });
+  try {
+    for (const form of forms) {
+      await deleteFolder('forms', form.id);
+      logger.info(`Files from form ${form.id} successfully removed.`);
+    }
+
+    const versions = forms.reduce((acc, form) => acc.concat(form.versions), []);
+    await Record.deleteMany({ form: { $in: forms } });
+    await Channel.deleteMany({ form: { $in: forms } });
+    await Version.deleteMany({ _id: { $in: versions } });
+  } catch (err) {
+    logger.error(`Deletion of forms failed: ${err.message}`);
+  }
 });
 
 schema.index(